+++ examples/simple-folder/main.tf
@@ -14,14 +14,33 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">= 0.12"
-}
-
 provider "google" {
-  version = "~> 3.31"
+  version = "~> 3.31.0"
 }
 
 provider "google-beta" {
-  version = "~> 3.31"
+  version = "~> 3.31.0"
+}
+
+provider "null" {
+  version = "~> 2.1"
+}
+
+provider "random" {
+  version = "~> 2.2"
+}
+
+/*************************************************
+  Bootstrap GCP Folder.
+*************************************************/
+
+module "seed_bootstrap" {
+  source               = "../.."
+  org_id               = var.org_id
+  parent_folder        = var.parent
+  billing_account      = var.billing_account
+  group_org_admins     = var.group_org_admins
+  group_billing_admins = var.group_billing_admins
+  default_region       = var.default_region
+  org_project_creators = var.org_project_creators
 }