+++ test/fixtures/simple-folder/main.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,14 +14,13 @@
  * limitations under the License.
  */
 
-terraform {
-  required_version = ">= 0.12"
-}
+module "simple" {
+  source = "../../../examples/simple-folder"
 
-provider "google" {
-  version = "~> 3.31"
-}
-
-provider "google-beta" {
-  version = "~> 3.31"
+  org_id               = var.org_id
+  parent               = var.parent_folder
+  billing_account      = var.billing_account
+  group_org_admins     = var.group_org_admins
+  group_billing_admins = var.group_billing_admins
+  default_region       = var.default_region
 }