+++ test/fixtures/simple-folder/versions.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019 Google LLC
+ * Copyright 2018 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,4 @@
 
 terraform {
   required_version = ">= 0.12"
-}
-
-provider "google" {
-  version = "~> 3.31"
-}
-
-provider "google-beta" {
-  version = "~> 3.31"
 }