+++ variables.tf
@@ -18,6 +18,12 @@
   Required variables
 *******************************************/
 
+variable "parent_folder" {
+  description = "GCP parent folder ID in the form folders/{id}"
+  default     = ""
+  type        = string
+}
+
 variable "org_id" {
   description = "GCP Organization ID"
   type        = string
@@ -133,4 +139,10 @@ variable "org_project_creators" {
   description = "Additional list of members to have project creator role accross the organization. Prefix of group: user: or serviceAccount: is required."
   type        = list(string)
   default     = []
+}
+
+variable "skip_gcloud_download" {
+  description = "Whether to skip downloading gcloud (assumes gcloud is already available outside the module)"
+  type        = bool
+  default     = true
 }